feat(avatar): update demo && snapshots
