+++ src/avatar/demos/size.vue
@@ -1,29 +1,83 @@
 <template>
   <div class="tdesign-demo-avatar">
-    <div class="tdesign-demo-block">
-      <t-avatar size="small">U</t-avatar>
-      <t-avatar size="medium">U</t-avatar>
-      <t-avatar size="large">U</t-avatar>
-      <t-avatar size="88px">U</t-avatar>
+    <div class="avatar-demo">
+      <t-avatar shape="circle" size="large" :icon="userIcon"></t-avatar>
+      <t-avatar shape="circle" size="medium" :icon="userIcon"></t-avatar>
+      <t-avatar shape="circle" size="small" :icon="userIcon"></t-avatar>
     </div>
-    <div class="tdesign-demo-block">
-      <t-avatar shape="round" size="small">U</t-avatar>
-      <t-avatar shape="round" size="medium">U</t-avatar>
-      <t-avatar shape="round" size="large">U</t-avatar>
-      <t-avatar shape="round" size="88px">U</t-avatar>
+    <div class="avatar-demo">
+      <t-avatar
+        shape="circle"
+        size="large"
+        image="https://tdesign.gtimg.com/mobile/demos/avatar_1.png"
+        alt="示例图片"
+        :badge-props="{ count: 10 }"
+      ></t-avatar>
+      <t-avatar
+        shape="circle"
+        size="medium"
+        image="https://tdesign.gtimg.com/mobile/demos/avatar_1.png"
+        alt="示例图片"
+        :badge-props="{ count: 10 }"
+      ></t-avatar>
+      <t-avatar
+        shape="circle"
+        size="small"
+        image="https://tdesign.gtimg.com/mobile/demos/avatar_1.png"
+        alt="示例图片"
+        :badge-props="{ dot: true }"
+      ></t-avatar>
+    </div>
+    <div class="avatar-demo">
+      <t-avatar
+        shape="round"
+        size="large"
+        image="https://tdesign.gtimg.com/mobile/demos/avatar_1.png"
+        alt="示例图片"
+        :badge-props="{ count: 10 }"
+      ></t-avatar>
+      <t-avatar
+        shape="round"
+        size="medium"
+        image="https://tdesign.gtimg.com/mobile/demos/avatar_1.png"
+        alt="示例图片"
+        :badge-props="{ count: 10 }"
+      ></t-avatar>
+      <t-avatar
+        shape="round"
+        size="small"
+        image="https://tdesign.gtimg.com/mobile/demos/avatar_1.png"
+        alt="示例图片"
+        :badge-props="{ dot: true }"
+      ></t-avatar>
+    </div>
+    <div class="avatar-demo">
+      <t-avatar shape="circle" size="large" class="custom">A</t-avatar>
+      <t-avatar shape="circle" size="medium" class="custom">A</t-avatar>
+      <t-avatar shape="circle" size="small" class="custom">A</t-avatar>
     </div>
   </div>
 </template>
 
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+import { h } from 'vue';
+import { UserIcon } from 'tdesign-icons-vue-next';
 
-<style lang="less" scoped>
-.tdesign-demo-avatar {
-  .tdesign-demo-avatar-block:first-child {
-    margin-bottom: 40px;
-  }
+const userIcon = () => h(UserIcon);
+</script>
+
+<style lang="less">
+.avatar-demo {
+  margin-left: 16px;
+  margin-bottom: 16px;
   .t-avatar {
-    margin-right: 40px;
+    margin-right: 20px;
+  }
+  .custom {
+    .t-avatar__inner {
+      background-color: #0052d9;
+      color: #fff;
+    }
   }
 }
 </style>
